List deposits list withdrawals (#412)
* ListDeposits-ListWithdrawals

* fixes

* fix for bug 404

* refactor after review

* refactor after review

* refactor after review

* refactor after review

+++ internal/core/application/operator_service.go
@@ -74,6 +74,11 @@ type OperatorService interface {
 		ctx context.Context,
 		req WithdrawMarketReq,
 	) ([]byte, error)
+	ListWithdrawals(
+		ctx context.Context,
+		accountIndex int,
+		page domain.Page,
+	) ([]domain.Withdrawal, error)
 	FeeAccountBalance(ctx context.Context) (int64, error)
 	ClaimMarketDeposit(
 		ctx context.Context,
@@ -84,6 +89,11 @@ type OperatorService interface {
 		ctx context.Context,
 		outpoints []TxOutpoint,
 	) error
+	ListDeposits(
+		ctx context.Context,
+		accountIndex int,
+		page domain.Page,
+	) ([]domain.Deposit, error)
 	ListMarket(
 		ctx context.Context,
 	) ([]MarketInfo, error)
@@ -868,8 +878,34 @@ func (o *operatorService) WithdrawMarketFunds(
 		log.Warn(err)
 	}
 
+	if err = o.repoManager.WithdrawalRepository().AddWithdrawal(
+		ctx,
+		domain.Withdrawal{
+			TxID:            txid,
+			AccountIndex:    accountIndex,
+			BaseAmount:      req.BalanceToWithdraw.BaseAmount,
+			QuoteAmount:     req.BalanceToWithdraw.QuoteAmount,
+			MillisatPerByte: req.MillisatPerByte,
+			Address:         req.Address,
+		},
+	); err != nil {
+		return nil, err
+	}
+
 	rawTx, _ := hex.DecodeString(txHex)
 	return rawTx, nil
+}
+
+func (o *operatorService) ListWithdrawals(
+	ctx context.Context,
+	accountIndex int,
+	page domain.Page,
+) ([]domain.Withdrawal, error) {
+	return o.repoManager.WithdrawalRepository().ListWithdrawalsForAccountIdAndPage(
+		ctx,
+		accountIndex,
+		page,
+	)
 }
 
 func (o *operatorService) FeeAccountBalance(ctx context.Context) (int64, error) {
@@ -1007,10 +1043,26 @@ func (o *operatorService) ClaimFeeDeposit(
 	return o.claimDeposit(ctx, infoPerAccount, outpoints, feeDeposit)
 }
 
+func (o *operatorService) ListDeposits(
+	ctx context.Context,
+	accountIndex int,
+	page domain.Page,
+) ([]domain.Deposit, error) {
+	return o.repoManager.DepositRepository().ListDepositsForAccountIdAndPage(
+		ctx,
+		accountIndex,
+		page,
+	)
+}
+
 func (o *operatorService) DropMarket(
 	ctx context.Context,
 	accountIndex int,
 ) error {
+	if accountIndex < domain.MarketAccountStart {
+		return ErrInvalidAccountIndex
+	}
+
 	return o.repoManager.MarketRepository().DeleteMarket(ctx, accountIndex)
 }
 
@@ -1099,6 +1151,7 @@ func (o *operatorService) claimDeposit(
 	// outpoints.
 	counter := make(map[int]int)
 	unspents := make([]domain.Unspent, len(outpoints), len(outpoints))
+	deposits := make([]domain.Deposit, 0, len(outpoints))
 	for i, v := range outpoints {
 		confirmed, err := o.explorerSvc.IsTransactionConfirmed(v.Hash)
 		if err != nil {
@@ -1146,6 +1199,14 @@ func (o *operatorService) claimDeposit(
 				Address:         info.Address,
 				Confirmed:       true,
 			}
+
+			deposits = append(deposits, domain.Deposit{
+				AccountIndex: info.AccountIndex,
+				TxID:         v.Hash,
+				VOut:         v.Index,
+				Asset:        unconfidential.AssetHash,
+				Value:        unconfidential.Value,
+			})
 		}
 	}
 
@@ -1166,6 +1227,13 @@ func (o *operatorService) claimDeposit(
 						return
 					}
 					log.Info("fee account funded. Trades can be served")
+
+					for _, v := range deposits {
+						err := o.repoManager.DepositRepository().AddDeposit(ctx, v)
+						if err != nil {
+							log.Error(err)
+						}
+					}
 				}
 			}()
